working

+++ Scripts/Book_Generator.py
@@ -1,11 +1,13 @@
 # Import libraries
 from PIL import ImageFont, ImageDraw, Image
 import shutil, os
+import numpy as np
+import cv2
 # Declare constant variables
 TEXT_FILE = 'Scripts/output.txt'
 IMAGES_LOCATION = 'Scripts/Output/'
 TEXT_FILE_CONTENTS = None
-LINES_PER_PAGE = 50
+LINES_PER_PAGE = 35
 CHARACTERS_PER_LINE = 40
 
 FONT_SIZE = 50
@@ -35,15 +37,11 @@ def clear_folders(folders):
         for filename in os.listdir(folder):
             file_path = os.path.join(folder, filename)
             try:
-                if os.path.isfile(file_path) or os.path.islink(file_path):
-                    os.unlink(file_path)
-                elif os.path.isdir(file_path):
-                    shutil.rmtree(file_path)
-            except Exception as e:
-                print('Failed to delete %s. Reason: %s' % (file_path, e))
-
-
+                if os.path.isfile(file_path) or os.path.islink(file_path): os.unlink(file_path)
+                elif os.path.isdir(file_path): shutil.rmtree(file_path)
+            except Exception as e: print('Failed to delete %s. Reason: %s' % (file_path, e))
 clear_folders([IMAGES_LOCATION])
+
 def add_margin(pil_img, top, right, bottom, left, color):
     width, height = pil_img.size
     new_width = width + right + left
@@ -57,19 +55,15 @@ with open(TEXT_FILE, 'r', encoding='utf-8') as file: TEXT_FILE_CONTENTS = file.r
 
 # Orginize text to fit on pages (start to end)
 word_ahead_char_count = 0
-text = TEXT_FILE_CONTENTS.split(' ')
+lines = TEXT_FILE_CONTENTS.split('\n')
+text = []
+for line in lines: text += line.split(' ')
 for word_index, word in enumerate(text):
+    if word == '':
+        orginized_text_file_contents.append(['\n'])
+        orginized_line = ''
+        char_count = 0
     char_count += len(word)
-    # if char_count >= CHARACTERS_PER_LINE:
-    #     try: word_ahead_char_count += char_count + len(text[word_index+1])
-    #     except IndexError: word_ahead_char_count = char_count
-    #     if word_ahead_char_count >= (CHARACTERS_PER_LINE + 10):
-    #         orginized_text_file_contents.append([orginized_line])
-    #         orginized_line = ''
-    #         char_count = 0
-    #     else:
-    #         orginized_line += (word + ' ')
-    # else:
     orginized_line += (word + ' ')
     if char_count >= CHARACTERS_PER_LINE:
         orginized_text_file_contents.append([orginized_line])
@@ -77,6 +71,7 @@ for word_index, word in enumerate(text):
         char_count = 0
 
 # Put lines in order of pages
+print(len(orginized_text_file_contents))
 page_text = []
 for _, line in enumerate(orginized_text_file_contents):
     line_count += 1
@@ -85,10 +80,13 @@ for _, line in enumerate(orginized_text_file_contents):
         page_layout_text.append(page_text)
         page_text = []
         line_count = 0
-
+page_text.clear()
+print(len(page_layout_text)*LINES_PER_PAGE)
 # How many coloumns are there
 NUM_OF_COLS = (len(page_layout_text))
-
+for i in orginized_text_file_contents[(len(page_layout_text)*LINES_PER_PAGE):]:
+    page_text.append(i)
+page_layout_text.append(page_text)
 if NUM_OF_COLS % 2 != 0:
     NUM_OF_COLS += 1
 
@@ -129,6 +127,7 @@ for page_num in file_names:
 files_with_text = file_names
 files_with_text.pop(0)
 files_with_text.pop(0)
+files_with_text.sort()
 for page_num in files_with_text:
     try:
         # print(page_layout_text[page_num])
@@ -136,17 +135,39 @@ for page_num in files_with_text:
         text = [i[0] for i in page_layout_text[page_num-1]]
         text = '\n'.join(text)
         img = Image.open(f"{IMAGES_LOCATION}{page_num}.png")
-        d = ImageDraw.Draw(img)
-        d.text((0, 0), text, fill=(0, 0, 0), font=FONT)
+        text_img = ImageDraw.Draw(img)
+        text_img.text((0, 0), text, fill=(0, 0, 0), font=FONT)
         img = add_margin(img, TOP_MARGIN, RIGHT_MARGIN, BOTTOM_MARGIN, LEFT_MARGIN, (255, 255, 255))
         img.save(f"{IMAGES_LOCATION}{page_num}.png", quality=95)
     except Exception as e: print(e)
 
-# Add margin
-
-
 # Add page numbers in footer
+flip_spots = True
+for page_order in ORDER_OF_PAGES:
+    for page in page_order:
+        img = Image.open(f"{IMAGES_LOCATION}{page}.png")
+        x, y = img.size
+        page_num_img = ImageDraw.Draw(img)
+        # Add margin
+        if flip_spots: 
+            page_num_img.text((100, 100), str(page), fill=(0, 0, 0), font=FONT)
+        else: 
+            page_num_img.text((x-150, 100), str(page), fill=(0, 0, 0), font=FONT)
+        flip_spots = not flip_spots
+        img.save(f"{IMAGES_LOCATION}{page}.png", quality=95)
 
 # Combine
+for page_order in ORDER_OF_PAGES:
+    
+    img1 = cv2.imread(f"{IMAGES_LOCATION}{page_order[0]}.png")
+
+    img2 = cv2.imread(f"{IMAGES_LOCATION}{page_order[1]}.png")
+
+    # img2.paste(img1, (0, 0), img1)
+    # img2.save(f'{page_order[0]} - {page_order[1]}.png',"PNG")
+    vis = np.concatenate((img1, img2), axis=1)
+    cv2.imwrite(f'{IMAGES_LOCATION}{page_order[0]} - {page_order[1]}.png', vis)
+    os.remove(f"{IMAGES_LOCATION}{page_order[0]}.png")
+    os.remove(f"{IMAGES_LOCATION}{page_order[1]}.png")
 
 # Save
